+++ packages/stargate/src/queries/utils.ts
@@ -1,4 +1,5 @@
 import { Bech32 } from "@cosmjs/encoding";
+import { Uint64 } from "@cosmjs/math";
 import { PageRequest } from "cosmjs-types/cosmos/base/query/v1beta1/pagination";
 import Long from "long";
 
@@ -13,6 +14,13 @@ export function toAccAddress(address: string): Uint8Array {
   return Bech32.decode(address).data;
 }
 
+/**
+ * If paginationKey is set, return a `PageRequest` with the given key.
+ * If paginationKey is unset, return `undefined`.
+ *
+ * Use this with a query response's pagination next key to
+ * request the next page.
+ */
 export function createPagination(paginationKey?: Uint8Array): PageRequest | undefined {
   return paginationKey
     ? {
@@ -35,4 +43,13 @@ export function createProtobufRpcClient(base: QueryClient): ProtobufRpcClient {
       return base.queryUnverified(path, data);
     },
   };
+}
+
+/**
+ * Takes a uint64 value as string, number, Long or Uint64 and returns an unsigned Long instance
+ * of it.
+ */
+export function longify(value: string | number | Long | Uint64): Long {
+  const checkedValue = Uint64.fromString(value.toString());
+  return Long.fromBytesBE([...checkedValue.toBytesBigEndian()], true);
 }